requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
pydantic
mysqlclient
python-dotenv